refactor!: merge style, bib, citation crates
