--- citation/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "csln-citation"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-schemars = "0.8.12"
-serde = "1.0.162"
-serde_derive = "1.0.162"
-serde_json = "1.0.96"
-serde_yaml = "0.9.21"
