+++ cli/Cargo.toml
@@ -18,8 +18,6 @@ path = "src/main.rs"
 clap = { version = "4.3.0", features = ["derive"] }
 schemars = "0.8.12"
 serde_json = "1.0.68"
-style = { path = "../style", package = "csln-style" }
-bibliography = { path = "../bibliography", package = "csln-bibliography" }
+csln = { path = "../csln", package = "csln" }
 processor = { path = "../processor", package = "csln-processor" }
-citation = { path = "../citation", package = "csln-citation" }
