+++ cli/src/main.rs
@@ -1,9 +1,9 @@
-use bibliography::HasFile;
-use bibliography::InputBibliography as Bibliography;
-use citation::Citation;
 use clap::Parser;
+use csln::bibliography::HasFile;
+use csln::bibliography::InputBibliography as Bibliography;
+use csln::citation::Citation;
+use csln::style::Style;
 use processor::Processor;
-use style::Style;
 
 #[derive(Parser, Default, Debug)]
 #[clap(author = "Bruce D'Arcus", version, about = "A CLI for CSLN")]
@@ -24,7 +24,7 @@ fn main() {
     let style: Style = Style::from_file(&opts.style);
     let bibliography: Bibliography = Bibliography::from_file(&opts.bibliography);
     let citations: Vec<Citation> = Vec::new();
-    let locale = style::locale::Locale::from_file(&opts.locale);
+    let locale = csln::style::locale::Locale::from_file(&opts.locale);
     let processor: Processor = Processor::new(style, bibliography, citations, locale);
     let rendered_refs = processor.render_references();
     println!("{}", serde_json::to_string_pretty(&rendered_refs).unwrap());
