+++ csln/Cargo.toml
@@ -1,18 +1,19 @@
 [package]
-name = "csln-bibliography"
+name = "csln"
 authors = ["Bruce D'Arcus <bdarcus@gmail.com>"]
 license = "MPL-2.0"
 version = "0.1.0"
 edition = "2021"
 
 [lib]
-name = "csln_bibliography" 
+name = "csln" 
 test = true
 doctest = true
 bench = true
 doc = true
 edition = "2021" 
 crate-type = ["lib"] 
+include = ["src/lib.rs", "src/style/*.rs", "src/bibliography/*rs"]
 
 [dependencies]
 schemars = { version = "0.8", features = ["url"] }
@@ -23,8 +24,10 @@ serde_yaml = "0.9"
 url = { version = "2.4.0", features = ["serde"] }
 edtf = { version = "0.2", features = ["chrono"] }
 chrono = { version = "0.4", features = ["unstable-locales"] }
-style = { path = "../style", package = "csln-style" }
-icu = { version = "1.2.0", features = ["icu_datetime_experimental"] }
-icu_testdata = { version = "1.2.0", features = ["icu_datetime_experimental"] }
-indexmap = { version = "2.0.0", features = ["std"] }
+unic-langid = { version = "0.9.1", features = ["serde"] }
+itertools = "0.11.0"
+rayon = "1.7.0"
+#icu = { version = "1.2.0", features = ["icu_datetime_experimental"] }
+#icu_testdata = { version = "1.2.0", features = ["icu_datetime_experimental"] }
+#indexmap = { version = "2.0.0", features = ["std"] }
