+++ csln/src/bibliography/.gitignore
