+++ processor/benches/proc_bench.rs
@@ -1,16 +1,15 @@
-use bibliography::HasFile;
-use bibliography::InputBibliography as Bibliography;
-use citation::Citation;
 use criterion::{criterion_group, criterion_main, Criterion};
+use csln::bibliography::HasFile;
+use csln::bibliography::InputBibliography as Bibliography;
+use csln::citation::Citation;
+use csln::style::Style;
 use csln_processor::Processor;
 use std::time::Duration;
-use style::Style;
 
 fn proc_benchmark(c: &mut Criterion) {
-    let style: Style = style::Style::from_file("examples/style.csl.yaml");
-    let bibliography: Bibliography =
-        bibliography::InputBibliography::from_file("examples/ex1.bib.yaml");
-    let locale = style::locale::Locale::from_file("locales/locale-en.yaml");
+    let style: Style = Style::from_file("examples/style.csl.yaml");
+    let bibliography: Bibliography = Bibliography::from_file("examples/ex1.bib.yaml");
+    let locale = csln::style::locale::Locale::from_file("locales/locale-en.yaml");
     let citations: Vec<Citation> = Vec::new();
     let processor: Processor = Processor::new(style, bibliography, citations, locale);
     c.bench_function("sorting references", |b| {
